+++ src/commands/addTodayAndStandupView.ts
@@ -1,0 +1,7 @@
+import { addToday } from './addToday';
+import { standupView } from './standupView';
+
+export async function addTodayAndStandupView() {
+    await addToday();
+    standupView();
+}